finished adding user criterias

+++ server.ts
@@ -140,16 +140,24 @@ app.get("/people", (req, res) => {
 });
 
 //retrieve all individuals whose most recent employment was in the specified company
+//Sends a 200 result with null information if user cannot see the company
 app.get("/people/:companyId", (req, res) => {
   try {
+    let userID = req.user.UserID;
+    let companyID = req.params.companyID;
     let be = new BackendProcessing();
-    let data = be.retrievePeopleViaCompany(req.params.companyId).then(results => {
-      if (!results) {
-        res.sendStatus(500);
-      } else {
-        res.json({ data: results });
-      }
-    });
+    if (!be.userSeesCompany(userID, companyID)) {
+      res.sendStatus(200);
+      return;
+    } else {
+      let data = be.retrievePeopleViaCompany(req.params.companyId).then(results => {
+        if (!results) {
+          res.sendStatus(500);
+        } else {
+          res.json({ data: results });
+        }
+      });
+    }
   } catch (error) {
     res.sendStatus(500);
   }
@@ -158,15 +166,23 @@ app.get("/people/:companyId", (req, res) => {
 app.put("/people", (req, res) => {
   //update person
   let be = new BackendProcessing();
-  //console.log("body: "+JSON.stringify(req.body))
   let person = req.body;
-  let update = be.update_person(person);
-  update.then(boolean => {
-    if (boolean) res.sendStatus(200);
-    else res.sendStatus(500);
-  });
-  update.catch(() => {
-    res.sendStatus(500);
+  let userID = req.user.UserID;
+  let fundID = person.FundID;
+  be.userCanChangeFund(userID, fundID).then(authorized => {
+    if (!authorized) {
+      console.error("user cannot update in the specified fund");
+      res.sendStatus(500);
+      return;
+    }
+    let update = be.update_person(person);
+    update.then(boolean => {
+      if (boolean) res.sendStatus(200);
+      else res.sendStatus(500);
+    });
+    update.catch(() => {
+      res.sendStatus(500);
+    });
   });
 });
 
@@ -174,30 +190,53 @@ app.post("/people", (req, res) => {
   //add person
   let be = new BackendProcessing();
   let person = req.body.newData;
-  let i = be.insert_person(person);
-  i.then(boolean => {
-    if (boolean) res.sendStatus(200);
-    else res.sendStatus(500);
-  });
-  i.catch(res.sendStatus(500));
+  let userID = req.user.UserID;
+  let fundID = person.FundID;
+  be.userCanChangeFund(userID, fundID)
+    .then(authorized => {
+      if (!authorized) {
+        console.error("user cannot add an individual to that fund");
+        res.sendStatus(500);
+        return;
+      }
+      let i = be.insert_person(person);
+      i.then(boolean => {
+        if (boolean) res.sendStatus(200);
+        else res.sendStatus(500);
+      });
+      i.catch(res.sendStatus(500));
+    })
+    .catch(() => {
+      res.sendStatus(500);
+    });
 });
 
 app.delete("/people/:id", (req, res) => {
   //delete person
   let be = new BackendProcessing();
   let person = req.params.id;
-  let d = be.delete_person(person);
-  d.then(boolean => {
-    if (boolean) res.sendStatus(200);
-    else res.sendStatus(500);
+  let userID = req.user.UserID;
+  let fundID = person.FundID;
+  be.userCanChangeFund(userID, fundID).then(authorized => {
+    if (!authorized) {
+      console.error("User cannot delete the individual");
+      res.sendStatus(500);
+      return;
+    }
+    let d = be.delete_person(person);
+    d.then(boolean => {
+      if (boolean) res.sendStatus(200);
+      else res.sendStatus(500);
+    });
+    d.catch(res.sendStatus(500));
   });
-  d.catch(res.sendStatus(500));
 });
 
 app.get("/company", (req, res) => {
   try {
     let be = new BackendProcessing();
-    let data = be.retrieveCompaniesFromDatabase().then(results => {
+    let userID = req.user.UserID;
+    let data = be.retrieveCompaniesFromDatabase(userID).then(results => {
       if (!results) {
         res.sendStatus(500);
       } else {
@@ -212,17 +251,25 @@ app.get("/company", (req, res) => {
 app.put("/company", (req, res) => {
   //update company
   let be = new BackendProcessing();
-  //console.log("body: "+JSON.stringify(req.body))
   let company = req.body;
-  let update = be.update_company(company);
-  update.then(boolean => {
-    if (boolean) {
-      console.log("company updated");
-      res.sendStatus(200);
-    } else res.sendStatus(500);
-  });
-  update.catch(() => {
-    res.sendStatus(500);
+  let userID = req.user.UserID;
+  let fundID = company.FundID;
+  be.userCanChangeFund(userID, fundID).then(authorized => {
+    if (!authorized) {
+      console.error("user is not authorized to change companies in that fund");
+      res.sendStatus(500);
+      return;
+    }
+    let update = be.update_company(company);
+    update.then(boolean => {
+      if (boolean) {
+        console.log("company updated");
+        res.sendStatus(200);
+      } else res.sendStatus(500);
+    });
+    update.catch(() => {
+      res.sendStatus(500);
+    });
   });
 });
 
@@ -230,33 +277,52 @@ app.post("/company", (req, res) => {
   //add company
   let be = new BackendProcessing();
   let company = req.body.newData;
-  let i = be.insert_company(company);
-  i.then(boolean => {
-    if (boolean) {
-      console.log("person added");
-      res.sendStatus(200);
-    } else res.sendStatus(500);
+  let userID = req.user.UserID;
+  let fundID = company.FundID;
+  be.userCanChangeFund(userID, fundID).then(authorized => {
+    if (!authorized) {
+      console.error("user cannot add a company to that fund");
+      res.sendStatus(500);
+      return;
+    }
+    let i = be.insert_company(company);
+    i.then(boolean => {
+      if (boolean) {
+        console.log("person added");
+        res.sendStatus(200);
+      } else res.sendStatus(500);
+    });
+    i.catch(res.sendStatus(500));
   });
-  i.catch(res.sendStatus(500));
 });
 
 app.delete("/company/:id", (req, res) => {
   //delete company
   let be = new BackendProcessing();
   let company = req.params.id;
-  let d = be.delete_company(company);
-  d.then(boolean => {
-    if (boolean) res.sendStatus(200);
-    else res.sendStatus(500);
+  let userID = req.user.UserID;
+  let fundID = company.FundID;
+  be.userCanChangeFund(userID, fundID).then(authorized => {
+    if (!authorized) {
+      console.error("user cannot delete a company from that fund");
+      res.sendStatus(500);
+      return;
+    }
+    let d = be.delete_company(company);
+    d.then(boolean => {
+      if (boolean) res.sendStatus(200);
+      else res.sendStatus(500);
+    });
+    d.catch(res.sendStatus(500));
   });
-  d.catch(res.sendStatus(500));
 });
 
 app.get("/funds", (req, res) => {
   //return type should be an Array of SideMenuFund objects as defined in App.tsx.
   try {
     let be = new BackendProcessing();
-    let results = be.retrieveFundsFromDatabase();
+    let userID = req.user.UserID;
+    let results = be.retrieveFundsFromDatabase(userID);
     results.then(funds => {
       res.json({ data: funds });
     });
@@ -282,25 +348,41 @@ app.put("/funds", (req, res) => {
   //req.body.fund.name is probably gonna be the thing that always change
   //Temp until function is implemented
   let be = new BackendProcessing();
-  console.log("Updated fund with id: " + req.body.fund.id + " to be called: " + req.body.fund.name);
-  let fund = req.body.fund;
-  be.update_fund(fund).then(result => {
-    if (result) res.sendStatus(200);
-    else res.sendStatus(500);
+  let userID = req.user.UserID;
+  be.userCanChangeFund(userID, req.body.fund.id).then(authorized => {
+    if (!authorized) {
+      console.error("User cannot update the fund");
+      res.sendStatus(500);
+      return;
+    }
+    console.log("Updated fund with id: " + req.body.fund.id + " to be called: " + req.body.fund.name);
+    let fund = req.body.fund;
+    be.update_fund(fund).then(result => {
+      if (result) res.sendStatus(200);
+      else res.sendStatus(500);
+    });
   });
 });
 
 app.get("/funds/:id", (req, res) => {
   try {
     let be = new BackendProcessing();
+    let userID = req.user.UserID;
     let fundID = req.params.id;
-    console.log("params: " + JSON.stringify(req.params));
-    console.log("type: " + typeof fundID);
-    console.log("string: " + fundID);
-    let data = be.retrieveCompaniesFromFund(fundID).then(results => {
-      if (results != null) {
-        res.json({ data: results });
-      } else res.sendStatus(500);
+    be.userSeesFund(userID, fundID).then(authorized => {
+      if (!authorized) {
+        console.error("User " + userID + " cannot see Fund " + fundID);
+        res.sendStatus(500);
+        return;
+      }
+      //console.log("params: " + JSON.stringify(req.params));
+      //console.log("type: " + typeof fundID);
+      //console.log("string: " + fundID);
+      let data = be.retrieveCompaniesFromFund(fundID).then(results => {
+        if (results != null) {
+          res.json({ data: results });
+        } else res.sendStatus(500);
+      });
     });
   } catch (error) {
     res.sendStatus(500);
@@ -311,12 +393,20 @@ app.get("/people/original/:companyId", (req, res) => {
   try {
     let be = new BackendProcessing();
     let companyID = req.params.id;
-    let data = be.retrievePeopleFromOriginalCompany(companyID).then(results => {
-      if (!results) {
+    let userID = req.user.UserID;
+    be.userSeesCompany(userID, companyID).then(authorized => {
+      if (!authorized) {
+        console.error("User " + userID + " cannot view company " + companyID);
         res.sendStatus(500);
-      } else {
-        res.json({ data: results });
+        return;
       }
+      let data = be.retrievePeopleFromOriginalCompany(companyID).then(results => {
+        if (!results) {
+          res.sendStatus(500);
+        } else {
+          res.json({ data: results });
+        }
+      });
     });
   } catch (error) {
     res.sendStatus(500);
@@ -326,12 +416,20 @@ app.get("/people/original/:companyId", (req, res) => {
 app.get("/history/:id", (req, res) => {
   let be = new BackendProcessing();
   let individualID = req.params.id;
-  let g = be.getHistoryOfIndividual(individualID);
-  g.then(results => {
-    res.json({ data: results });
-  });
-  g.catch(error => {
-    res.sendStatus(500);
+  let userID = req.user.UserID;
+  be.userCanSeeIndividual(userID, individualID).then(authorized => {
+    if (!authorized) {
+      console.error("User cannot see history of individual " + individualID);
+      res.sendStatus(500);
+      return;
+    }
+    let g = be.getHistoryOfIndividual(individualID);
+    g.then(results => {
+      res.json({ data: results });
+    });
+    g.catch(error => {
+      res.sendStatus(500);
+    });
   });
 });
 
@@ -340,16 +438,23 @@ app.put("/history", (req, res) => {
   let history = req.body;
   let userID = req.user.UserID;
   let individual = req.employee;
-  let ins = be.insertHistory(history, individual, userID);
-  ins.then(result => {
-    if (result) {
-      res.sendStatus(200);
-    } else {
+  be.userCanChangeIndividual(userID, individual.individualID).then(authorized => {
+    if (!authorized) {
+      console.error("User cannot add history to individual " + individual.IndividualID);
       res.sendStatus(500);
+      return;
     }
-  });
-  ins.catch(error => {
-    res.sendStatus(500);
+    let ins = be.insertHistory(history, individual, userID);
+    ins.then(result => {
+      if (result) {
+        res.sendStatus(200);
+      } else {
+        res.sendStatus(500);
+      }
+    });
+    ins.catch(error => {
+      res.sendStatus(500);
+    });
   });
 });
 
@@ -358,16 +463,23 @@ app.post("/history/:id", (req, res) => {
   let history = req.body;
   let individual = req.employee;
   let userID = req.user.UserID;
-  let upd = be.updateHistory(history, individual, userID);
-  upd.then(result => {
-    if (result) {
-      res.sendStatus(200);
-    } else {
+  be.userCanChangeIndividual(userID, individual.individualID).then(authorized => {
+    if (!authorized) {
+      console.error("User cannot edit history of individual " + individual.IndividualID);
       res.sendStatus(500);
+      return;
     }
-  });
-  upd.catch(error => {
-    res.sendStatus(500);
+    let upd = be.updateHistory(history, individual, userID);
+    upd.then(result => {
+      if (result) {
+        res.sendStatus(200);
+      } else {
+        res.sendStatus(500);
+      }
+    });
+    upd.catch(error => {
+      res.sendStatus(500);
+    });
   });
 });
